+++ subduction_wasm_bootstrap/src/lib.rs
@@ -23,7 +23,10 @@ extern crate alloc;
 #[cfg(target_arch = "wasm32")]
 pub mod js_logger;
 
-#[cfg(target_arch = "wasm32")]
+// Only the basic console stack (`install_basic_tracing`) uses these at the
+// crate root, and it is itself gated on `wasm-tracing`. The rich stack has its
+// own imports in `rich.rs`.
+#[cfg(all(target_arch = "wasm32", feature = "wasm-tracing"))]
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
 #[cfg(all(target_arch = "wasm32", feature = "wasm-tracing"))]
@@ -77,49 +80,7 @@ pub fn init_basic() {
 // ---------------------------------------------------------------------------
 
 #[cfg(target_arch = "wasm32")]
-mod rich {
-    use std::sync::OnceLock;
-
-    use tracing_subscriber::{
-        Registry, filter::LevelFilter, layer::SubscriberExt, reload, util::SubscriberInitExt,
-    };
-
-    use crate::js_logger::JsCallbackLayer;
-
-    /// Global handle for dynamically reloading the log-level filter.
-    static RELOAD_HANDLE: OnceLock<reload::Handle<LevelFilter, Registry>> = OnceLock::new();
-
-    pub(crate) fn init(initial_level: LevelFilter) {
-        let (filter, reload_handle) = reload::Layer::new(initial_level);
-
-        let registry = tracing_subscriber::registry()
-            .with(filter)
-            .with(JsCallbackLayer);
-
-        #[cfg(feature = "wasm-tracing")]
-        {
-            // WasmLayer accepts everything; the reloadable LevelFilter (added
-            // first, above) controls what passes through.
-            let mut config =
-                wasm_tracing::WasmLayerConfig::new().with_max_level(tracing::Level::TRACE);
-            config.use_console_methods = true;
-            registry.with(wasm_tracing::WasmLayer::new(config)).init();
-        }
-
-        #[cfg(not(feature = "wasm-tracing"))]
-        registry.init();
-
-        RELOAD_HANDLE.set(reload_handle).ok();
-    }
-
-    pub(crate) fn set_level(
-        level: LevelFilter,
-    ) -> Result<(), std::boxed::Box<dyn std::error::Error + Send + Sync>> {
-        let handle = RELOAD_HANDLE.get().ok_or("tracing not initialized")?;
-        handle.modify(|filter| *filter = level)?;
-        Ok(())
-    }
-}
+pub mod rich;
 
 /// Install the panic hook and the rich tracing stack (reloadable level filter +
 /// browser-console layer + JS-callback layer) at `initial_level`.
